Moves types and interfaces to "./types" file
